requirements.txt
streamlit 
firebase
firestore
vecto-sdk
streamlit-option-menu
python-dotenv
openai
Flask
requests
flask
streamlit_extras